Split cli-tests.cpp into per-domain registration files, add CLAUDE.md
Break the 1629-line monolithic test file into 12 focused files:
- cli-tests-main.cpp (entry point)
- cli-core-tests.cpp (broker, HTTP, serial, peer, button, light, display, command)
- fdn-protocol-tests.cpp (protocol parsing, FDN game, device extensions)
- signal-echo-tests.cpp (Signal Echo gameplay + difficulty)
- fdn-integration-tests.cpp (integration, complete, progress, challenge, color profile)
- firewall-decrypt-tests.cpp (address, lifecycle, managed, config, state names)
- progression-e2e-tests.cpp (E2E, re-encounter, color context, bounty role)
- ghost-runner-reg-tests.cpp, spike-vector-reg-tests.cpp, cipher-path-reg-tests.cpp,
  exploit-seq-reg-tests.cpp, breach-defense-reg-tests.cpp (per-minigame registrations)

All 344 CLI tests pass. Eliminates the #1 merge conflict source for
multi-agent parallel development.

Also adds CLAUDE.md to repo root with Multi-VM Agent Conventions section
and updated test file convention docs for the new per-domain split.
