+++ test/test_cli/breach-defense-reg-tests.cpp
@@ -1,0 +1,83 @@
+//
+// Breach Defense Tests — Registration file for Breach Defense minigame
+//
+
+#include <gtest/gtest.h>
+
+#include "breach-defense-tests.hpp"
+
+// ============================================
+// BREACH DEFENSE TESTS
+// ============================================
+
+TEST_F(BreachDefenseTestSuite, EasyConfigPresets) {
+    breachDefenseEasyConfigPresets(this);
+}
+
+TEST_F(BreachDefenseTestSuite, HardConfigPresets) {
+    breachDefenseHardConfigPresets(this);
+}
+
+TEST_F(BreachDefenseTestSuite, IntroResetsSession) {
+    breachDefenseIntroResetsSession(this);
+}
+
+TEST_F(BreachDefenseTestSuite, IntroTransitionsToShow) {
+    breachDefenseIntroTransitionsToShow(this);
+}
+
+TEST_F(BreachDefenseTestSuite, ShowDisplaysThreatInfo) {
+    breachDefenseShowDisplaysThreatInfo(this);
+}
+
+TEST_F(BreachDefenseTestSuite, ShowTransitionsToGameplay) {
+    breachDefenseShowTransitionsToGameplay(this);
+}
+
+TEST_F(BreachDefenseTestSuite, ThreatAdvancesWithTime) {
+    breachDefenseThreatAdvancesWithTime(this);
+}
+
+TEST_F(BreachDefenseTestSuite, CorrectBlock) {
+    breachDefenseCorrectBlock(this);
+}
+
+TEST_F(BreachDefenseTestSuite, MissedThreat) {
+    breachDefenseMissedThreat(this);
+}
+
+TEST_F(BreachDefenseTestSuite, ShieldMovesUpDown) {
+    breachDefenseShieldMovesUpDown(this);
+}
+
+TEST_F(BreachDefenseTestSuite, EvaluateRoutesToNextThreat) {
+    breachDefenseEvaluateRoutesToNextThreat(this);
+}
+
+TEST_F(BreachDefenseTestSuite, EvaluateRoutesToWin) {
+    breachDefenseEvaluateRoutesToWin(this);
+}
+
+TEST_F(BreachDefenseTestSuite, EvaluateRoutesToLose) {
+    breachDefenseEvaluateRoutesToLose(this);
+}
+
+TEST_F(BreachDefenseTestSuite, WinSetsOutcome) {
+    breachDefenseWinSetsOutcome(this);
+}
+
+TEST_F(BreachDefenseTestSuite, LoseSetsOutcome) {
+    breachDefenseLoseSetsOutcome(this);
+}
+
+TEST_F(BreachDefenseTestSuite, StandaloneLoopsToIntro) {
+    breachDefenseStandaloneLoopsToIntro(this);
+}
+
+TEST_F(BreachDefenseTestSuite, StateNamesResolve) {
+    breachDefenseStateNamesResolve(this);
+}
+
+TEST_F(BreachDefenseManagedTestSuite, ManagedModeReturns) {
+    breachDefenseManagedModeReturns(this);
+}
