+++ test/test_cli/exploit-seq-reg-tests.cpp
@@ -1,0 +1,87 @@
+//
+// Exploit Sequencer Tests — Registration file for Exploit Sequencer minigame
+//
+
+#include <gtest/gtest.h>
+
+#include "exploit-sequencer-tests.hpp"
+
+// ============================================
+// EXPLOIT SEQUENCER TESTS
+// ============================================
+
+TEST_F(ExploitSequencerTestSuite, EasyConfigPresets) {
+    exploitSeqEasyConfigPresets(this);
+}
+
+TEST_F(ExploitSequencerTestSuite, HardConfigPresets) {
+    exploitSeqHardConfigPresets(this);
+}
+
+TEST_F(ExploitSequencerTestSuite, IntroResetsSession) {
+    exploitSeqIntroResetsSession(this);
+}
+
+TEST_F(ExploitSequencerTestSuite, IntroTransitionsToShow) {
+    exploitSeqIntroTransitionsToShow(this);
+}
+
+TEST_F(ExploitSequencerTestSuite, ShowDisplaysSequenceInfo) {
+    exploitSeqShowDisplaysSequenceInfo(this);
+}
+
+TEST_F(ExploitSequencerTestSuite, ShowTransitionsToGameplay) {
+    exploitSeqShowTransitionsToGameplay(this);
+}
+
+TEST_F(ExploitSequencerTestSuite, SymbolAdvancesWithTime) {
+    exploitSeqSymbolAdvancesWithTime(this);
+}
+
+TEST_F(ExploitSequencerTestSuite, CorrectTimingHit) {
+    exploitSeqCorrectTimingHit(this);
+}
+
+TEST_F(ExploitSequencerTestSuite, WrongTimingMiss) {
+    exploitSeqWrongTimingMiss(this);
+}
+
+TEST_F(ExploitSequencerTestSuite, TimeoutCountsFail) {
+    exploitSeqTimeoutCountsFail(this);
+}
+
+TEST_F(ExploitSequencerTestSuite, EvaluateRoutesToNextExploit) {
+    exploitSeqEvaluateRoutesToNextExploit(this);
+}
+
+TEST_F(ExploitSequencerTestSuite, EvaluateAdvancesSequence) {
+    exploitSeqEvaluateAdvancesSequence(this);
+}
+
+TEST_F(ExploitSequencerTestSuite, EvaluateRoutesToWin) {
+    exploitSeqEvaluateRoutesToWin(this);
+}
+
+TEST_F(ExploitSequencerTestSuite, EvaluateRoutesToLose) {
+    exploitSeqEvaluateRoutesToLose(this);
+}
+
+TEST_F(ExploitSequencerTestSuite, WinSetsOutcome) {
+    exploitSeqWinSetsOutcome(this);
+}
+
+TEST_F(ExploitSequencerTestSuite, LoseSetsOutcome) {
+    exploitSeqLoseSetsOutcome(this);
+}
+
+TEST_F(ExploitSequencerTestSuite, StandaloneLoopsToIntro) {
+    exploitSeqStandaloneLoopsToIntro(this);
+}
+
+TEST_F(ExploitSequencerTestSuite, StateNamesResolve) {
+    exploitSeqStateNamesResolve(this);
+}
+
+TEST_F(ExploitSequencerManagedTestSuite, ManagedModeReturns) {
+    exploitSeqManagedModeReturns(this);
+}
